backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Remetra backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.0,<5.0.0",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "ruff>=0.14.11",
    "scalar-fastapi>=1.6.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
commit = [
    "pre-commit>=4.5.1",
]
dev = [
    "httpx>=0.28.1",
    "pytest-cov>=7.0.0",
]
